+++ sdk/cosmos/azure-cosmos/azure/cosmos/container.py
@@ -497,6 +497,10 @@ class ContainerProxy:  # pylint: disable=too-many-public-methods
             LATEST_VERSION: Query latest items from 'start_time' or 'continuation' token.
             ALL_VERSIONS_AND_DELETES: Query all versions and deleted items from either `start_time='Now'`
             or 'continuation' token.
+            All versions and deletes responses contain ``current`` and ``metadata`` keys. The **provisional**
+            ``previous`` key can contain the previous item version when previous images are enabled for the container.
+            To request access to this provisional feature, see `previous image opt-in
+            <https://aka.ms/cosmosdb-change-feed-deletes>`_.
         :paramtype mode: Literal["LatestVersion", "AllVersionsAndDeletes"]
         :keyword Sequence[str] excluded_locations: Excluded locations to be skipped from preferred locations. The locations
             in this list are specified as the names of the azure Cosmos locations like, 'West US', 'East US' and so on.
@@ -547,6 +551,10 @@ class ContainerProxy:  # pylint: disable=too-many-public-methods
             LATEST_VERSION: Query latest items from 'start_time' or 'continuation' token.
             ALL_VERSIONS_AND_DELETES: Query all versions and deleted items from either `start_time='Now'`
             or 'continuation' token.
+            All versions and deletes responses contain ``current`` and ``metadata`` keys. The **provisional**
+            ``previous`` key can contain the previous item version when previous images are enabled for the container.
+            To request access to this provisional feature, see `previous image opt-in
+            <https://aka.ms/cosmosdb-change-feed-deletes>`_.
         :paramtype mode: Literal["LatestVersion", "AllVersionsAndDeletes"]
         :keyword Sequence[str] excluded_locations: Excluded locations to be skipped from preferred locations. The locations
             in this list are specified as the names of the azure Cosmos locations like, 'West US', 'East US' and so on.
@@ -632,6 +640,10 @@ class ContainerProxy:  # pylint: disable=too-many-public-methods
             LATEST_VERSION: Query latest items from 'start_time' or 'continuation' token.
             ALL_VERSIONS_AND_DELETES: Query all versions and deleted items from either `start_time='Now'`
             or 'continuation' token.
+            All versions and deletes responses contain ``current`` and ``metadata`` keys. The **provisional**
+            ``previous`` key can contain the previous item version when previous images are enabled for the container.
+            To request access to this provisional feature, see `previous image opt-in
+            <https://aka.ms/cosmosdb-change-feed-deletes>`_.
         :paramtype mode: Literal["LatestVersion", "AllVersionsAndDeletes"]
         :keyword Sequence[str] excluded_locations: Excluded locations to be skipped from preferred locations. The locations
             in this list are specified as the names of the azure Cosmos locations like, 'West US', 'East US' and so on.
@@ -680,6 +692,10 @@ class ContainerProxy:  # pylint: disable=too-many-public-methods
             LATEST_VERSION: Query latest items from 'start_time' or 'continuation' token.
             ALL_VERSIONS_AND_DELETES: Query all versions and deleted items from either `start_time='Now'`
             or 'continuation' token.
+            All versions and deletes responses contain ``current`` and ``metadata`` keys. The **provisional**
+            ``previous`` key can contain the previous item version when previous images are enabled for the container.
+            To request access to this provisional feature, see `previous image opt-in
+            <https://aka.ms/cosmosdb-change-feed-deletes>`_.
         :paramtype mode: Literal["LatestVersion", "AllVersionsAndDeletes"]
         :keyword Sequence[str] excluded_locations: Excluded locations to be skipped from preferred locations. The locations
             in this list are specified as the names of the azure Cosmos locations like, 'West US', 'East US' and so on.