+++ sdk/cosmos/azure-cosmos/samples/change_feed_management_async.py
@@ -54,6 +54,19 @@ async def clean_up(container):
         # Deleting the current item
         await container.delete_item(item, partition_key=item['address']['state'])
 
+
+def print_all_versions_and_deletes_change(change):
+    metadata = change.get('metadata', {})
+    print('Operation: {}'.format(metadata.get('operationType')))
+    print('Current item: {}'.format(change.get('current')))
+
+    # `previous` is provisional and is returned only when previous images are enabled for the container.
+    # Request preview access at https://aka.ms/cosmosdb-change-feed-deletes.
+    previous = change.get('previous')
+    if previous is not None:
+        print('Previous item: {}'.format(previous))
+
+
 async def read_change_feed(container):
     print('\nReading Change Feed from the beginning\n')
 
@@ -134,7 +147,7 @@ async def read_change_feed_with_all_versions_and_delete_mode(container):
     # This initial call was made to store a point in time in a 'continuation' token
     response_iterator = container.query_items_change_feed(mode="AllVersionsAndDeletes")
     async for doc in response_iterator:
-        print(doc)
+        print_all_versions_and_deletes_change(doc)
     continuation_token = container.client_connection.last_response_headers['etag']
 
     # Read all change feed with 'AllVersionsAndDeletes' mode after create items from a continuation
@@ -142,13 +155,13 @@ async def read_change_feed_with_all_versions_and_delete_mode(container):
     await create_items(container, 10, 'OR')
     response_iterator = container.query_items_change_feed(mode="AllVersionsAndDeletes", continuation=continuation_token)
     async for doc in response_iterator:
-        print(doc)
+        print_all_versions_and_deletes_change(doc)
 
     # Read all change feed with 'AllVersionsAndDeletes' mode after delete items from a continuation
     await clean_up(container)
     response_iterator = container.query_items_change_feed(mode="AllVersionsAndDeletes", continuation=continuation_token)
     async for doc in response_iterator:
-        print(doc)
+        print_all_versions_and_deletes_change(doc)
 
 async def read_change_feed_with_all_versions_and_delete_mode_with_partition_key(container):
     print('\nReading Change Feed with AllVersionsAndDeletes mode from the partition key\n')
@@ -157,7 +170,7 @@ async def read_change_feed_with_all_versions_and_delete_mode_with_partition_key(
     # This initial call was made to store a point in time and 'partition_key' in a 'continuation' token
     response_iterator = container.query_items_change_feed(mode="AllVersionsAndDeletes", partition_key="CA")
     async for doc in response_iterator:
-        print(doc)
+        print_all_versions_and_deletes_change(doc)
     continuation_token = container.client_connection.last_response_headers['etag']
 
     await create_items(container, 10, 'CA')
@@ -166,7 +179,7 @@ async def read_change_feed_with_all_versions_and_delete_mode_with_partition_key(
     # Should only print the created items with 'CA' partition key value
     response_iterator = container.query_items_change_feed(mode='AllVersionsAndDeletes', continuation=continuation_token)
     async for doc in response_iterator:
-        print(doc)
+        print_all_versions_and_deletes_change(doc)
     continuation_token = container.client_connection.last_response_headers['etag']
 
     await clean_up(container)
@@ -174,7 +187,7 @@ async def read_change_feed_with_all_versions_and_delete_mode_with_partition_key(
     # Should only print the deleted items with 'CA' partition key value
     response_iterator = container.query_items_change_feed(mode='AllVersionsAndDeletes', continuation=continuation_token)
     async for doc in response_iterator:
-        print(doc)
+        print_all_versions_and_deletes_change(doc)
 
 async def run_sample():
     async with CosmosClient(HOST, MASTER_KEY) as client: